Addressed Several Assignments
Additions:
* ShowAllMovies: Created a new page to show all movies in the database.
I added a table, to be populated the same way all of the other tables
are.

* In the footer, I added a link to "Contact Us"

Modifications:
* Edited the space between the navigation bar and the first div,
headline, or image, so when the page is smaller, the navigation bar does
not cut it off.

* Changed the page names in aspx and aspx.cs files:
- Advanced_Search is now AdvancedSearch.
- Rewards_Points is now RewardsPoints.
- Search_Results is now SearchResults.
- User_Profile is now UserProfile
- TermsandConditions is now TermsAndConditions

* Cleaned up the footer portion of pages Login, Manage, Register by
adjusting the spacing.

* Updated the cover image url to /Images/CrowdTagMovieTheater.png

* Updated the image in that file location as the image was not the most
recent version.

* Site.Master: Changed the tabs in the navigation bar.  When the tabs
take too much space and the screen is narrow, the navigation bar became
three rows.  In the current format, even when a user is logged in, it
only uses two rows.
- Also updated the links in the navigation bar to the new page names.